+++ tsdown.config.ts
@@ -1,12 +1,8 @@
-import { type Options, defineConfig } from 'tsdown';
+import { defineConfig, type UserConfig } from 'tsdown/config';
 
 export default defineConfig({
-  clean: true,
-  dts: true,
-  entry: ['src/index.ts'],
+  fixedExtension: true,
   format: ['esm', 'cjs'],
-  platform: 'node',
-  removeNodeProtocol: true,
-  sourcemap: !process.env.IS_RELEASE,
-  target: 'node12'
-}) as Options;
+  nodeProtocol: 'strip',
+  sourcemap: !process.env.IS_RELEASE
+}) as UserConfig;